fix(grid): correct SattrT type from i16 to i32
The Rust type alias SattrT was incorrectly defined as i16, but C's
sattr_T is int32_t (4 bytes). This size mismatch caused memory
corruption when writing to grid_attrs from Rust:

- Rust writes at offset 80 * 2 = 160 bytes
- C reads from offset 80 * 4 = 320 bytes

This led to the assertion failure 'attrs[i] >= 0' in ui_comp_raw_line
when starting nvim without a file argument (intro screen).

+++ src/nvim-rs/grid/src/lib.rs
@@ -542,8 +542,8 @@ pub extern "C" fn rs_schar_get_first_codepoint(sc: ScharT) -> c_int {
 // Phase 30: Grid Line Content Functions
 // =============================================================================
 
-/// Type alias for screen attribute (matches C's `sattr_T` which is `int16_t`).
-type SattrT = i16;
+/// Type alias for screen attribute (matches C's `sattr_T` which is `int32_t`).
+type SattrT = i32;
 
 /// Type alias for column number (matches C's `colnr_T` which is `int32_t`).
 type ColnrT = i32;
